+++ openedx_events/event_bus/avro/types.py
@@ -1,11 +1,14 @@
 """A mapping of python types to the Avro type that we want to use make valid avro schema."""
-PYTHON_TYPE_TO_AVRO_MAPPING = {
-    None: "null",
+SIMPLE_PYTHON_TYPE_TO_AVRO_MAPPING = {
     bool: "boolean",
     int: "long",
     float: "double",
     bytes: "bytes",
     str: "string",
+}
+PYTHON_TYPE_TO_AVRO_MAPPING = {
+    **SIMPLE_PYTHON_TYPE_TO_AVRO_MAPPING,
+    None: "null",
     dict: "record",
     list: "array",
 }